FavouritesDbHelper: add plugin interface
For nextcloud, we just need three touch points, add, delete and load
the points, so add them via a single static plugin (so we can only
support a single plugin at once ... if there were a reason to support
more, we'd convert this to an event fire with notification
registrations interface).  Because the plugin interface may take time
to execute if it has a web backend, an asynchronous execution
structure is introduced.  Because of the asynchronous component, we
assume that the initial points we get from the local files is what we
work with and the eventual completion of the back end may or may not
update them.  For this reason, the initial set of points is the one
presented and the web updated copy is available some time later
... although it is securely committed to local storage after it
becomes available.  Likewise point addition, deletion and editing all
occurs with immediate local saving and eventual plugin update.

This strategy allows for fast initialization and prevents the UI
hanging while the web service is queried.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>