FavouritePoint: Add extensions to handle external map sources
External map sources need to be able to have the ability to set
additional properties.  The way we can do this is to place them into
the extensions piece of the waypoint, but this necessitates access to
the extensions as a key value store and the ability to set and get
individual elements, so expose the APIs.

To get this to work, we have to alter the FavouritesDbHelper.java file
to save all the extensions it finds instead of only color and HIDDEN.
As an internal tidy up, FavouritePoint now treats the color and
visible attributes of FavouritePoint as extensions.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>