FavouritesDbHelper: add last modified time to gpx points
In order to tell whether a remote point should be updated or not, we
add a <modified> tag to the extensions giving the modification time in
seconds past the epoch (this is how nextcloud stores its modifications
so it gives us something directly comparable).

For GPX files that do not have this tag, we simply add it at the
current time.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>