Nexcloud: make plugin reconcile dates by modification time
Since the nextcloud modification time is immutable, we must always
rely on that, so every time a point is added or updated on a nextcloud
remote, we always save the new modification time obtained from the
nextcloud server in the local file.  This means that two points that
haven't been modified should have the same local and remote
modification time.  Reconciliation now becomes simple, simply take the
point with the latest modification time as definitive.

Rename detection is slightly subtle: we already store the nextcloud id
in the extensions, so we do lookups by id instead of name to build the
array.  However, there will be cases where id isn't present (offline
creation of the point or ineffective update) in those cases we do
lookup by name and create the remote if it doesn't exist.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>